Complete splaytree with count

+++ tree/splaytree/splay.h
@@ -1,12 +1,12 @@
 typedef struct SplayNode SplayNode;
 
 struct SplayNode {
-  i64 key;
-  i64 countLeft, countRight;
+  int key;
+  int countLeft, countRight;
   SplayNode *left, *right;
 };
 
-SplayNode *NewSplayNode(i64 key);
-SplayNode *SplaySearch(SplayNode*, i64 key);
-SplayNode *SplayInsert(SplayNode*, i64 key);
-SplayNode *SplayDelete(SplayNode*, i64 key);
+SplayNode *NewSplayNode(int key);
+SplayNode *SplaySearch(SplayNode*, int key);
+SplayNode *SplayInsert(SplayNode*, int key);
+SplayNode *SplayDelete(SplayNode*, int key);
